refactor: group stdin object store helpers into StdinUtils
Move the stdin scheme constant, location helpers, and in-memory object
store construction (plus their tests) out of object_storage.rs into a
dedicated object_storage/stdin.rs submodule, encapsulated as StdinUtils.
